requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
e2b-code-interpreter>=0.1.0
together>=0.2.8 